Abstract out properties
